initial button move
