+++ apps/i18n/codebridge/en_us.json
@@ -29,9 +29,11 @@
   "runToContinue": "You must run your code before continuing.",
   "runToFinish": "You must run your code before finishing.",
   "runToSubmit": "You must run your code before submitting.",
+  "stopValidation": "Stop validation",
   "systemCodeError": "We're sorry, we hit an issue running your program. Please refresh the page and try again.",
   "testName": "Test Name",
   "testsDidNotPass": "Tests did not all pass.",
+  "validate": "Validate",
   "validationNotYetPassedContinue": "You must pass tests before continuing.",
   "validationNotYetPassedFinish": "You must pass tests before finishing.",
   "validationNotYetPassedSubmit": "You must pass tests before submitting.",
