--- apps/src/codebridge/ControlButtons/control-buttons.module.scss
@@ -1,31 +1,0 @@
-@import 'color.scss';
-
-.controlButtonsContainer {
-  grid-area: control-buttons;
-  display: grid;
-  grid-template-columns: 1fr auto 1fr;
-  padding: 8px 0;
-  background-color: $neutral_dark80;
-}
-
-.centerButton {
-  grid-column-start: 2;
-  display: flex;
-}
-
-.runButton {
-  margin-right: 8px;
-}
-
-.navigationButton {
-  margin-left: auto;
-  margin-right: 8px;
-  font-size: 16px;
-  display: flex;
-}
-
-.disabledInfoIcon {
-  padding: 8px;
-  color: $white;
-  cursor: 'pointer'
-}
