+++ apps/src/lab2/lab2Redux.ts
@@ -299,7 +299,7 @@ export const isReadOnlyWorkspace = (state: RootState) => {
   const readonlyPredictLevel = isReadonlyPredictLevel(state);
   const hasSubmitted = getCurrentLevel(state)?.status === LevelStatus.submitted;
   const isRunningAndReadonly =
-    (state.lab2System.isRunning || state.lab2System.isTesting) &&
+    (state.lab2System.isRunning || state.lab2System.isValidating) &&
     shouldBeReadonlyWhileRunning(state);
 
   return (
