+++ apps/src/pythonlab/PythonlabView.tsx
@@ -42,20 +42,19 @@ const defaultProject: ProjectSources = {
 
 const labeledGridLayouts = {
   horizontal: {
-    gridLayoutRows: '2fr 1fr 48px',
+    gridLayoutRows: '2fr 1fr ',
     gridLayoutColumns: '300px minmax(0, 1fr)',
     gridLayout: `
   "info-panel workspace"
   "file-browser console"
-  "file-browser control-buttons"`,
+  `,
   },
   vertical: {
-    gridLayoutRows: '1fr 1fr 48px',
+    gridLayoutRows: '1fr 1fr',
     gridLayoutColumns: '300px minmax(0, 1fr) minmax(0, 1fr)',
     gridLayout: `
     "info-panel workspace console"
-    "file-browser workspace console"
-    "file-browser control-buttons control-buttons"`,
+    "file-browser workspace console"`,
   },
 };
 const defaultConfig: ConfigType = {
